requirements.txt
moviepy>=1.0.3
python-dotenv>=0.19.0
openai>=0.27.0
ffmpeg-python>=0.2.0
pydub>=0.25.1
argparse>=1.4.0
